test/AvaliadorTest.java

import exemploleilaotdd.Avaliador;
import modelo.Lance;
import modelo.Leilao;
import modelo.Usuario;
import org.junit.Assert;
import org.junit.Test;

/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
/**
 *
 * @author MAICON
 */
public class AvaliadorTest {

    @Test
    public void deveEncontrarMaiorLance() {
        Usuario joao = new Usuario("Joao");
        Usuario jose = new Usuario("José");
        Usuario maria = new Usuario("Maria");
        Leilao leilao = new Leilao("Playstation 3 Novo");
        
        leilao.propoe(new Lance(maria, 250.0));
        leilao.propoe(new Lance(joao, 300.0));
        leilao.propoe(new Lance(jose, 400.0));
        
        Avaliador leiloeiro = new Avaliador();
        leiloeiro.avalia(leilao);
        
        double maiorEsperado = 400;
        Assert.assertEquals(maiorEsperado, leiloeiro.getMaiorLance(), 0);
    }
    
    @Test
    public void deveEncontrarMenorLance() {
        Usuario joao = new Usuario("Joao");
        Usuario jose = new Usuario("José");
        Usuario maria = new Usuario("Maria");
        Leilao leilao = new Leilao("Playstation 3 Novo");
        
        leilao.propoe(new Lance(maria, 300.0));
        leilao.propoe(new Lance(joao, 400.0));
        leilao.propoe(new Lance(jose, 250.0));
        
        Avaliador leiloeiro = new Avaliador();
        leiloeiro.avalia(leilao);
        
        double menorEsperado = 250;
        Assert.assertEquals(menorEsperado, leiloeiro.getMenorLance(), 0);
    }    
}
